requirements.txt
streamlit
geopy
requests
google-cloud-vision
google-cloud-texttospeech
openai
python-dotenv